add docker

+++ main.py
@@ -7,6 +7,8 @@ import subprocess
 import datetime
 import re
 import random
+import os
+import sys
 
 main = Blueprint('main', __name__)
 admin='cssp'
@@ -22,8 +24,8 @@ def train():
     gpu_info = subprocess.run(args=['nvidia-smi','--query-gpu=index,name','--format=csv,noheader'],stdout=subprocess.PIPE,universal_newlines=True)
     gpu_string = gpu_info.stdout
     gpu_list = re.split('\n|, ',gpu_string.rstrip())
-    test=["1","3090"]
-    gpu_list+=test
+    # test=["1","3090"]
+    # gpu_list+=test
     gpu_sql = db.session.query(Gpu_status).all()
     for data in gpu_sql:
         gpu_list.remove(data.index)
@@ -64,19 +66,38 @@ def train_post():
         if framework == None or version == None or gpu == [] or gpu_left==True :
             flash('Please choose the right framework, version and GPU!')
             #return redirect(url_for('main.train'))
+ 
         else:
-            start_time = datetime.datetime.now().strftime("%Y-%m-%d_%H:%M:%S")
+            start_time = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
             docker_name = current_user.name +"_" + framework +"_"+ version +"_" + start_time
             gpu_usage = ', '.join(gpu_list[1::2])
-            port = str(random.randint(1000,9999))
+            gpu_usage_index = ','.join(gpu_list[::2])
+            port = str(random.randint(1619,9999))
             full_port = "140.113.149.172:"+ port
             tasks = task(name=current_user.name, docker_name=docker_name, start_time=full_port,gpu=gpu_usage)
-
-            db.session.add(tasks)
-            db.session.commit()
-            for i in range (len(gpu_list)//2):
-                db.session.add(Gpu_status(index=gpu_list[2*i],gpu_name=gpu_list[2*i+1]))
+            folder = "/media/data/"+current_user.name+"/"
+            models = "/media/data/"+current_user.name+"/models/"
+            data = "/media/data/"+current_user.name+"/data/"
+            os.makedirs(folder, exist_ok=True)
+            os.makedirs(models, exist_ok=True)
+            os.makedirs(data, exist_ok=True)
+            command="docker run -d -p "+ port +":22 --name "+ docker_name +" --gpus device=" \
+            + gpu_usage_index + "  -v /media/data/"+current_user.name +"/:/workspace -w /workspace cssp618/environment:cssp_" \
+            + framework +"_"+ version 
+            
+            # p1 = subprocess.run(command,shell=True)
+            p1 = subprocess.run(args=['docker','run','-d','-p', port+':22','--name',docker_name, \
+            '--gpus','device='+gpu_usage_index,'-v','/media/data/'+current_user.name+'/:/root/', \
+            '-w','/root/','cssp618/environment:cssp_'+framework +"_"+ version+'_ws'])
+            if p1.returncode == 0 :
+                db.session.add(tasks)
                 db.session.commit()
+                for i in range (len(gpu_list)//2):
+                    db.session.add(Gpu_status(index=gpu_list[2*i],gpu_name=gpu_list[2*i+1]))
+                    db.session.commit()
+            else:
+                print(p1.stderr)
+                flash("Please find admin to solve the problem !")
 
         return redirect(url_for('main.train'))
 
@@ -84,10 +105,14 @@ def train_post():
         gpu_name = request.form.get('gpu')
         docker_name = request.form.get('docker_name')
         gpu_list = re.split(', ',gpu_name)
-        task.query.filter_by(docker_name=docker_name).delete()
-        for name in gpu_list:
-            Gpu_status.query.filter_by(gpu_name=name).delete()
-        db.session.commit()
+        p2 = subprocess.run(args=['docker','rm','-f',docker_name])
+        if p2.returncode == 0 :
+            task.query.filter_by(docker_name=docker_name).delete()
+            for name in gpu_list:
+                Gpu_status.query.filter_by(gpu_name=name).delete()
+            db.session.commit()
+        else:
+            pass
         # print (docker_name)
         return redirect(url_for('main.train'))
